[networkx] Update to 3.6.1
Closes: #15116
