+++ stubs/networkx/networkx/algorithms/bipartite/matrix.pyi
@@ -16,4 +16,11 @@ def biadjacency_matrix(
     format="csr",
 ): ...  # Return is a complex union of scipy classes depending on the format param
 @_dispatchable
-def from_biadjacency_matrix(A, create_using: Graph[_Node] | None = None, edge_attribute: str = "weight"): ...
+def from_biadjacency_matrix(
+    A,
+    create_using: Graph[_Node] | None = None,
+    edge_attribute: str = "weight",
+    *,
+    row_order: Iterable[Incomplete] | None = None,
+    column_order: Iterable[Incomplete] | None = None,
+): ...
