+++ stubs/networkx/networkx/algorithms/community/__init__.pyi
@@ -1,9 +1,9 @@
 from networkx.algorithms.community.asyn_fluid import *
+from networkx.algorithms.community.bipartitions import *
 from networkx.algorithms.community.centrality import *
 from networkx.algorithms.community.community_utils import *
 from networkx.algorithms.community.divisive import *
 from networkx.algorithms.community.kclique import *
-from networkx.algorithms.community.kernighan_lin import *
 from networkx.algorithms.community.label_propagation import *
 from networkx.algorithms.community.leiden import *
 from networkx.algorithms.community.local import *
