+++ stubs/networkx/networkx/classes/graph.pyi
@@ -43,7 +43,7 @@ class Graph(Collection[_Node]):
     def to_undirected_class(self) -> type[Graph[_Node]]: ...
     # @_dispatchable adds `backend` argument, but this decorated is unsupported constructor type here
     # and __init__() ignores this argument
-    def __new__(cls, incoming_graph_data: _Data[_Node] | None = None, *, backend=None, **attr: Any) -> Self: ...
+    def __new__(cls, *args, backend=None, **kwargs) -> Self: ...
     def __init__(self, incoming_graph_data: _Data[_Node] | None = None, **attr: Any) -> None: ...  # attr: key=value pairs
     @cached_property
     def adj(self) -> AdjacencyView[_Node, _Node, dict[str, Any]]: ...
