+++ stubs/networkx/networkx/classes/multigraph.pyi
@@ -24,7 +24,7 @@ class MultiGraph(Graph[_Node]):
     def to_undirected_class(self) -> type[MultiGraph[_Node]]: ...
     # @_dispatchable adds `backend` argument, but this decorated is unsupported constructor type here
     # and __init__() ignores this argument
-    def __new__(cls, incoming_graph_data=None, multigraph_input: bool | None = None, *, backend=None, **attr: Any) -> Self: ...
+    def __new__(cls, *args, backend=None, **kwargs) -> Self: ...
     def __init__(self, incoming_graph_data=None, multigraph_input: bool | None = None, **attr: Any) -> None: ...
     @cached_property
     def adj(self) -> MultiAdjacencyView[_Node, _Node, dict[str, Any]]: ...  # data can be any type
